Refactor GameLevel classes to extend GameLevelCsPathIdentity for shared behavior
